requirements.txt
langchain
faiss-cpu
python-dotenv
tiktoken
streamlit
groq
sentence-transformers
transformers
torch 
reportlab
reportlab python-pptx